refactor: rename runtime package and centralize shared types
