+++ README.md
@@ -2,7 +2,10 @@
 
 This repository is a monorepo containing:
 
-- `@noxigui/runtime` – the core runtime implementation.
+- `@noxigui/core` – foundational types, layout engine and UI elements.
+- `@noxigui/parser` – XML parser that produces runtime UI trees.
+- `@noxigui/runtime-core` – minimal runtime wiring parser output to a renderer.
+- `@noxigui/renderer-pixi` – PIXI-based renderer implementation.
 - `@noxigui/playground` – a Vite playground for experimenting with the runtime.
 
 Run `pnpm dev` to start the playground or `pnpm build` to build all packages.
