+++ packages/core/src/common/geometry.ts
@@ -1,0 +1,2 @@
+export type Size = { width: number; height: number };
+export type Rect = { x: number; y: number; width: number; height: number };
