+++ packages/core/src/elements/Image.ts
@@ -1,4 +1,5 @@
-import { UIElement, type Size, type Rect } from './UIElement.js';
+import { UIElement } from './UIElement.js';
+import type { Size, Rect } from '../common/geometry.js';
 
 export class Image extends UIElement {
   hAlign: 'Left' | 'Center' | 'Right' = 'Left';
