+++ packages/core/src/elements/Text.ts
@@ -1,4 +1,5 @@
-import { UIElement, type Size, type Rect } from './UIElement.js';
+import { UIElement } from './UIElement.js';
+import type { Size, Rect } from '../common/geometry.js';
 
 export abstract class Text extends UIElement {
   content: string;
