+++ packages/parser/src/parsers/BorderParser.ts
@@ -1,8 +1,8 @@
-import { BorderPanel } from '../../../runtime/src/elements/BorderPanel.js';
-import { applyGridAttachedProps, parseSizeAttrs, parseColor, parseMargin, applyMargin } from '../../../runtime/src/helpers.js';
+import { BorderPanel } from '../../../runtime-core/src/elements/BorderPanel.js';
+import { applyGridAttachedProps, parseSizeAttrs, parseColor, parseMargin, applyMargin } from '../../../runtime-core/src/helpers.js';
 import type { ElementParser } from './ElementParser.js';
 import type { Parser } from '../Parser.js';
-import type { UIElement } from '../../../core/src/index.js';
+import type { UIElement } from '@noxigui/core';
 import type * as PIXI from 'pixi.js';
 
 /** Parser for `<Border>` elements. */
