+++ packages/runtime-core/src/runtime.ts
@@ -1,8 +1,7 @@
 import { Parser } from '../../parser/src/Parser.js';
 import { Grid } from './elements/Grid.js';
-import { UIElement } from '../../core/src/index.js';
-import type { Size } from '../../core/src/index.js';
-import type { Renderer } from './renderer.js';
+import { UIElement } from '@noxigui/core';
+import type { Size, Renderer } from '@noxigui/core';
 
 export const RuntimeInstance = {
   create(xml: string, renderer: Renderer) {
