+++ packages/runtime-core/tsconfig.json
@@ -5,13 +5,7 @@
     "declaration": true,
     "declarationMap": true,
     "allowImportingTsExtensions": false,
-    "erasableSyntaxOnly": false,
-    "baseUrl": "src",
-    "paths": {
-      "@noxigui/core": ["../core/src/index.ts"],
-      "@noxigui/core/*": ["../core/src/*"]
-    },
-    "moduleResolution": "node"
+    "erasableSyntaxOnly": false
   },
   "include": ["src"]
 }
